requirements.txt
albumentations==2.0.8
lmdb==1.7.3
matplotlib==3.10.5
networkx==3.5
numpy==2.2.6
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
pandas==2.3.1
Pillow==11.3.0
python_Levenshtein==0.27.1
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.7.1
six==1.17.0
#torch==2.7.1+cu128
#torchvision==0.22.1+cu128
tqdm==4.67.1
#ultralytics==8.3.174 #这是yolo的包
